+++ src/report/index.ts
@@ -1,5 +1,6 @@
 export { buildJsonReport } from "./json-reporter.js";
 export { renderMarkdownReport } from "./markdown-reporter.js";
+export { renderSarifReport } from "./sarif-reporter.js";
 export { renderPrComment } from "./pr-comment-renderer.js";
 export { riskLevelForScore } from "./report-utils.js";
 export type {